requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
httpx
python-telegram-bot
openai
python-dotenv
requests
fastmcp